+++ android/src/main/kotlin/co/paystack/flutterpaystack/MethodCallHandlerImpl.kt
@@ -2,14 +2,13 @@ package co.paystack.flutterpaystack
 
 import android.annotation.SuppressLint
 import android.app.Activity
-import android.os.Build
 import android.provider.Settings
 import io.flutter.plugin.common.BinaryMessenger
 import io.flutter.plugin.common.MethodCall
 import io.flutter.plugin.common.MethodChannel
 import io.flutter.plugin.common.MethodChannel.MethodCallHandler
 
-class MethodCallHandlerImpl(messenger: BinaryMessenger?, private val activity: Activity?) : MethodCallHandler {
+class MethodCallHandlerImpl(messenger: BinaryMessenger, private val activity: Activity?) : MethodCallHandler {
     private var channel: MethodChannel? = null
     private var authDelegate: AuthDelegate? = null