+++ lib/src/widgets/common/extensions.dart
@@ -1,0 +1,7 @@
+import 'package:flutter/material.dart';
+
+extension Context on BuildContext {
+  TextTheme textTheme() => Theme.of(this).textTheme;
+
+  ColorScheme colorScheme() => Theme.of(this).colorScheme;
+}