+++ lib/src/widgets/error_widget.dart
@@ -51,7 +51,6 @@ class ErrorWidget extends StatelessWidget {
               text!,
               textAlign: TextAlign.center,
               style: const TextStyle(
-                color: Colors.black54,
                 fontWeight: FontWeight.w500,
                 fontSize: 14.0,
               ),