+++ lib/src/widgets/input/base_field.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
+import 'package:flutter_paystack/src/widgets/common/extensions.dart';
 
 class BaseTextField extends StatelessWidget {
   final Widget? suffix;
@@ -50,7 +51,7 @@ class BaseTextField extends StatelessWidget {
             borderSide: const BorderSide(color: Colors.grey, width: 0.5)),
         focusedBorder: OutlineInputBorder(
             borderSide:
-                BorderSide(color: Theme.of(context).accentColor, width: 1.0)),
+                BorderSide(color: context.colorScheme().secondary, width: 1.0)),
         hintText: hintText,
       ),
     );