+++ lib/src/widgets/otp_widget.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_paystack/src/widgets/base_widget.dart';
+import 'package:flutter_paystack/src/widgets/common/extensions.dart';
 import 'package:flutter_paystack/src/widgets/custom_dialog.dart';
 import 'package:flutter_paystack/src/widgets/input/otp_field.dart';
 
@@ -39,9 +40,9 @@ class _OtpWidgetState extends BaseState<OtpWidget> {
                 new Text(
                   widget.message!,
                   textAlign: TextAlign.center,
-                  style: const TextStyle(
+                  style:  TextStyle(
                     fontWeight: FontWeight.w500,
-                    color: Colors.black87,
+                    color: context.textTheme().headline6?.color,
                     fontSize: 15.0,
                   ),
                 ),
@@ -50,7 +51,7 @@ class _OtpWidgetState extends BaseState<OtpWidget> {
                   padding: EdgeInsets.symmetric(horizontal: 10),
                   child: new OtpField(
                     onSaved: (String? value) => _otp = value,
-                    borderColor: Theme.of(context).accentColor,
+                    borderColor: context.colorScheme().secondary,
                   ),
                 ),
                 heightBox,