+++ lib/src/widgets/sucessful_widget.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_paystack/src/common/utils.dart';
 import 'package:flutter_paystack/src/widgets/animated_widget.dart';
+import 'package:flutter_paystack/src/widgets/common/extensions.dart';
 
 class SuccessfulWidget extends StatefulWidget {
   final int amount;
@@ -54,7 +55,7 @@ class _SuccessfulWidgetState extends State<SuccessfulWidget>
             }
           });
 
-    WidgetsBinding.instance!.addPostFrameCallback((_) => _startCountdown());
+    WidgetsBinding.instance.addPostFrameCallback((_) => _startCountdown());
   }
 
   @override
@@ -67,7 +68,7 @@ class _SuccessfulWidgetState extends State<SuccessfulWidget>
 
   @override
   Widget build(BuildContext context) {
-    final accentColor = Theme.of(context).accentColor;
+    final sceondaryColor = context.colorScheme().secondary;
     return new Container(
       child: new CustomAnimatedWidget(
         controller: _mainController,
@@ -77,15 +78,15 @@ class _SuccessfulWidgetState extends State<SuccessfulWidget>
             sizedBox,
             new Image.asset(
               'assets/images/successful.png',
-              color: accentColor,
+              color: sceondaryColor,
               width: 50.0,
               package: 'flutter_paystack',
             ),
             sizedBox,
-            const Text(
+            Text(
               'Payment Successful',
-              style: const TextStyle(
-                color: Colors.black87,
+              style: TextStyle(
+                color: context.textTheme().headline6?.color,
                 fontWeight: FontWeight.w500,
                 fontSize: 16.0,
               ),
@@ -96,8 +97,8 @@ class _SuccessfulWidgetState extends State<SuccessfulWidget>
             widget.amount.isNegative
                 ? new Container()
                 : new Text('You paid ${Utils.formatAmount(widget.amount)}',
-                    style: const TextStyle(
-                      color: Colors.black54,
+                    style: TextStyle(
+                      color: context.textTheme().headline6?.color,
                       fontWeight: FontWeight.normal,
                       fontSize: 14.0,
                     )),
@@ -107,7 +108,7 @@ class _SuccessfulWidgetState extends State<SuccessfulWidget>
               child: new Text(
                 _countdownAnim.value.toString(),
                 style: TextStyle(
-                    color: accentColor,
+                    color: sceondaryColor,
                     fontWeight: FontWeight.bold,
                     fontSize: 25.0),
               ),