+++ test/src/common/utils_test.dart
@@ -2,7 +2,6 @@ import 'package:flutter/services.dart';
 import 'package:flutter_paystack/src/common/my_strings.dart';
 import 'package:flutter_paystack/src/common/utils.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:matcher/matcher.dart';
 
 void main() {
   const MethodChannel channel =